% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_selection_function.R
\name{plot_selection}
\alias{plot_selection}
\title{Randomly place circular plots within a given (sf) area and with a given radius}
\usage{
plot_selection(n, inner_polygon, radius, outer_polygon = NULL)
}
\arguments{
\item{n}{Sample size of plots to be included.}

\item{inner_polygon}{Area that the centre of plots should be randomised within}

\item{radius}{Radius of the requested plot(s) in metres}

\item{outer_polygon}{Area that the entirety of the plot should be randomised within. NB - This can be the same as the inner polygon}
}
\description{
Randomly place circular plots, with a given radius, within an sf object, so that the plots will not overlap.
}
\examples{
plot_selection(n=450, inner_polygon=Soton, radius=11.4, outer_polygon=Soton)
}
\keyword{circular}
\keyword{plots}
